consolidate impl
